// src/api/generated/model/solarInverterDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Chargee Ampere
 * API v2
 * OpenAPI spec version: 2.0.0
 */
import type { SolarInverterInfoDto } from './solarInverterInfoDto';
import type { SolarInverterDtoLastProductionState } from './solarInverterDtoLastProductionState';

export interface SolarInverterDto {
  /** Solar inverter identifier */
  identifier: string;
  /** Solar inverter information */
  info: SolarInverterInfoDto;
  /**
   * Last production state
   * @nullable
   */
  lastProductionState?: SolarInverterDtoLastProductionState;
  /** Created at timestamp */
  createdAt: string;
  /**
   * Updated at timestamp
   * @nullable
   */
  updatedAt: string | null;
}